id input handler

+++ Events/Lesson/Modal.js
@@ -26,7 +26,7 @@ async function LessonModal(interaction) {
 			let nextId = await Database.getLessonNextQuestionId(interaction.user.id, interaction.channel.id, lesson, questionId);
 
 			if (nextId === 'there_is_no_more_questions') {
-				await interaction.channel.send(lessonFinishedEmbed(interaction.user.id, interaction.channel.id))
+				await interaction.channel.send(await lessonFinishedEmbed(interaction.user.id, interaction.channel.id))
                 await interaction.channel.send({ content: `*Šis channel tiks izdzēsts pēc 1 minutes automātiski*` })
                 await Database.addUserCoins(interaction.user.id, await Database.getActiveLessonRewardCountTotal(interaction.user.id, interaction.channel.id))
 				await Database.addToUserLessonPoints(interaction.user.id, lesson, await Database.getActiveLessonRewardCountTotal(interaction.user.id, interaction.channel.id))
@@ -34,7 +34,9 @@ async function LessonModal(interaction) {
                 await Database.addToUserHistoryALesson(interaction.user.id, await Database.getActiveLessonByChannel(interaction.channel.id))
                 await Database.deleteActiveLesson(interaction.channel.id);
                 setTimeout(async () => {
-                    await interaction.channel.delete();
+					if (interaction.channel != null) {
+						await interaction.channel.delete();
+					}
                 }, 60000);
 			} else {
 				await interaction.channel.send(await createQuestionEmbed(lesson, nextId, interaction.user.id, false, false, ''));
